refactor: move loupe to actions
