--- site/pages/reference/account-kit/smart-contracts/functions/accountLoupeV08Actions.mdx
@@ -1,37 +1,0 @@
----
-# This file is autogenerated
-title: accountLoupeV08Actions
-description: Overview of the accountLoupeV08Actions method
----
-
-# accountLoupeV08Actions
-
-Provides a set of actions for account loupe operations using the specified client.
-NOTE: this is already added to the client when using any of the Modular Account Clients.
-
-## Import
-
-```ts
-import { accountLoupeV08Actions } from "@account-kit/smart-contracts";
-```
-
-## Usage
-
-```ts
-import { accountLoupeActions } from "@account-kit/smart-contracts";
-import { createSmartAccountClient } from "@aa-sdk/core";
-
-const client = createSmartAccountClient(...).extend(accountLoupeV08Actions);
-```
-
-## Parameters
-
-### client
-
-`Client<TTransport, TChain, TAccount>`
-the client to be used for executing the account loupe actions
-
-## Returns
-
-`AccountLoupeActions<TAccount>`
-an object containing account loupe actions like `getExecutionFunctionConfig`, `getExecutionHooks`, `getPreValidationHooks`, and `getInstalledPlugins`
